Adds a mobile menu for larger screens, small nits
